backend/requirements/base.txt
asgiref==3.4.1
autopep8==1.6.0
backports.zoneinfo==0.2.1
Django==4.0.1
django-filter==21.1
djangorestframework==3.13.1
flake8==4.0.1
importlib-metadata==4.10.0
isort==5.10.1
Markdown==3.3.6
mccabe==0.6.1
Pillow==9.0.0
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
python-dotenv==0.19.2
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
zipp==3.7.0
